+++ ReleaseNotes/release_notes.md
@@ -8,6 +8,8 @@
   1. add more contructors to `Ostap::Math::Polynomial` class 
   1. unify the key-function for case-insensitive (and noi underscores) dictionaries 
   1. improve `ostap.fitting.roostats` and corresponding test 
+  1. more polishing for `ostap.fitting.roostats` and corresponding test 
+
 
 ## Backward incompatible: